[feat] : Session 생성 서비스 로직

+++ src/main/java/nextstep/courses/domain/session/SessionRepository.java
@@ -1,7 +1,7 @@
 package nextstep.courses.domain.session;
 
 public interface SessionRepository {
-    int save(Session session, Long courseId);
+    long save(Session session, Long courseId);
 
     Session findById(Long id);
 }
